requirements.txt
psycopg2-binary~=2.9.10
apache-airflow~=3.0.3